pyproject.toml
[project]
name = "wsi2brick"
version = "0.1.0"
description = "A simple converter, generationg (lego) brick mosaics from whole slide images"
authors = [
    {name = "Daniel Hieber", email = "github@teddowyn.de"},
]
dependencies = [
    "openslide-python>=1.3.1",
    "numpy>=1.26.4",
    "scikit-image>=0.23.2",
    "tqdm>=4.66.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
testing = [
    "pytest>=8.2.0",
    "flake8>=7.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pdm.scripts]
test = "pdm run -v pytest tests"
lint = "pdm run -v flake8 --ignore=E124,E127,E128,E701,E731 --max-line-length 120 src tests"
# run prior to tests
pre_test = "pdm install"